http: add request content streaming
Currently, the seastar::httpd::request object has a content field, into
which the entire body of the request is read. Some requests may have
very large bodies, and it is inefficient and/or impossible to read the
entire body up-front into a string. This patch adds a content_stream
field which is an input_stream<char> wrapped around the connection
input_stream, from which the handler can read the entire content at its
own pace. The content field is still availalbe for handlers unable to
use the stream, but new applications using the http server should make
handlers that use the new field. To use the stream feature, the server
should be set up with a set_content_streaming(true) call.

Fixes #637

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>