http: add "Transfer-Encoding: chunked" handling
The length of a body of some requests may remain unknown
until after beginning sending it. For such requests
the body may be transferred in chunks.
This patch implements decoding requests that use the
"Transfer-Encoding: chunked" header. The Content-Length
header is then ignored and the handler is presented with
the decoded body.

Fixes #776

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>