iostream: add reading/skipping all bytes in an input_stream
If an application knows that it will need all bytes from a buffer,
instead of reading chunks of it and combining it, it can now use the
read_all() method to get all bytes in one net::packet.
Some applications may know that they're going to receive small amounts
of data in a stream. For them it may be more convenient to work on a
single buffer instead of a net::packet. The read_all_contiguous()
method provides such a buffer.
An app (e.g. an http handler) may need to exhaust an entire input_stream
(e.g. to enable processing the next http request on the same connection).
This can now be achieved by using the skip_all() method.

Refs: scylladb/scylla#7213

Signed-off-by: Wojciech Mitros <wojciech.mitros@scylladb.com>